fix(ttl): upgrade epoch and TTL types from u32 to u64
The insertion epoch counter (AtomicU32/u32) wrapped around after ~4.29B
inserts, causing expired entries to magically 'un-expire'. The cross-
shard current_epoch() using naive max() also broke on wrap-around.

Upgraded all epoch and TTL types from u32 to u64 across the entire
codebase (raw.rs, sync.rs, sharded.rs, lib.rs, examples). u64 will
not overflow within any reasonable timeframe (~584 years at 1B ops/sec).

BREAKING CHANGE: Public API methods set_ttl(), get_ttl(),
current_epoch(), and insert_ttl() now use u64 instead of u32.

Fixes: TTL Epoch u32 Overflow (Audit PR #4)

+++ examples/all_scenarios.rs
@@ -33,7 +33,7 @@ fn run_ttl_expiry() {
     map.insert_ttl(
         "immortal_key".to_string(),
         "lives_forever".to_string(),
-        u32::MAX,
+        u64::MAX,
     );
 
     for i in 0..11 {
